requirements.txt
streamlit>=1.32.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
slack-sdk>=3.27.0
python-dotenv>=1.0.0
